requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
pydantic>=2.9.0
python-jose[cryptography]=3.4.0
bcrypt>=4.0.0
python-multipart>=0.0.18
email-validator>=2.1.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

#Safety audit
pip_audit>=2.10.0
bandit>=1.9.2